requirements.txt
boto3
requests
progressbar2
